requirements.txt
# Ksher AgentAI 智能工作台 — Python依赖

# Web框架
streamlit>=1.40.0

# AI模型客户端
anthropic>=0.45.0
openai>=1.60.0

# 数据可视化
plotly>=5.24.0

# 数据处理
pandas>=2.2.0

# PPT生成
python-pptx>=1.0.0

# 图片处理
Pillow>=11.0.0

# 环境变量
python-dotenv>=1.0.0
